% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLesionFeatures.R
\name{getLesionFeatures}
\alias{getLesionFeatures}
\title{Feature computation to use in LINDA}
\usage{
getLesionFeatures(img, template, bmask, sigma = 2, verbose = TRUE)
}
\arguments{
\item{img}{subject scan (antsImage)}

\item{template}{template (antsImage)}

\item{bmask}{subject brain mask (antsImage)}

\item{sigma}{(default=2) smoothing parameter}

\item{verbose}{(default=TRUE) whether to display
console messages}
}
\value{
list of six features (antsImages)
}
\description{
Feature computation to use in LINDA
}
