% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linda.R
\name{linda.plot}
\alias{linda.plot}
\title{Plot linda results}
\usage{
linda.plot(
  linda.obj,
  variables.plot,
  titles = NULL,
  alpha = 0.05,
  lfc.cut = 1,
  legend = FALSE,
  directory = NULL,
  width = 11,
  height = 8
)
}
\arguments{
\item{linda.obj}{return from function \code{linda}.}

\item{variables.plot}{vector; variables whose results are to be plotted. For example, suppose the return
value \code{variables} is equal to \code{('x1', 'x2', 'x3b', 'x3c', 'x1:x2')}, then one could set \code{variables.plot = c('x3b', 'x1:x2')}.}

\item{titles}{vector; titles of the effect size plot and volcano plot for each variable in \code{variables.plot}.
The default is NULL. If NULL, the titles will be set as \code{variables.plot}.}

\item{alpha}{a real value between 0 and 1; cutoff for \code{padj}.}

\item{lfc.cut}{a positive value; cutoff for \code{log2FoldChange}.}

\item{legend}{TRUE or FALSE; whether to show the legends of the effect size plot and volcano plot.}

\item{directory}{character; the directory to save the figures, e.g., \code{getwd()}. The default is NULL. If NULL, figures will not be saved.}

\item{width}{the width of the graphics region in inches. See R function \code{pdf}.}

\item{height}{the height of the graphics region in inches. See R function \code{pdf}.}
}
\value{
A list of \code{ggplot2} objects.
\item{plot.lfc}{a list of effect size plots. Each plot corresponds to one variable in \code{variables.plot}.}
\item{plot.volcano}{a list of volcano plots. Each plot corresponds to one variable in \code{variables.plot}.}
}
\description{
The function plots the effect size plot and volcano plot based on the output from \code{linda}.
}
\examples{

#install package "phyloseq" for importing "smokers" dataset
ind <- smokers$meta$AIRWAYSITE == 'Throat'
otu.tab <- as.data.frame(smokers$otu[, ind])
meta <- cbind.data.frame(Smoke = factor(smokers$meta$SMOKER[ind]),
                         Sex = factor(smokers$meta$SEX[ind]),
                         Site = factor(smokers$meta$SIDEOFBODY[ind]),
                         SubjectID = factor(smokers$meta$HOST_SUBJECT_ID[ind]))
ind1 <- which(meta$Site == 'Left')
res.left <- linda(otu.tab[, ind1], meta[ind1, ], formula = '~Smoke+Sex', alpha = 0.1,
                  prev.cut = 0.1, lib.cut = 1000, winsor.quan = 0.97)
ind2 <- which(meta$Site == 'Right')
res.right <- linda(otu.tab[, ind2], meta[ind2, ], formula = '~Smoke+Sex', alpha = 0.1,
                  prev.cut = 0.1, lib.cut = 1000, winsor.quan = 0.97)
rownames(res.left$output[[1]])[which(res.left$output[[1]]$reject)]
rownames(res.right$output[[1]])[which(res.right$output[[1]]$reject)]

linda.obj <- linda(otu.tab, meta, formula = '~Smoke+Sex+(1|SubjectID)', alpha = 0.1,
                   prev.cut = 0.1, lib.cut = 1000, winsor.quan = 0.97)
linda.plot(linda.obj, c('Smokey', 'Sexmale'),
           titles = c('Smoke: n v.s. y', 'Sex: female v.s. male'), alpha = 0.1, lfc.cut = 1,
           legend = TRUE, directory = NULL, width = 11, height = 8)

}
\references{
Huijuan Zhou, Kejun He, Jun Chen, and Xianyang Zhang. LinDA: Linear Models for Differential Abundance
Analysis of Microbiome Compositional Data.
}
\author{
Huijuan Zhou \email{huijuanzhou2019@gmail.com}
Jun Chen \email{Chen.Jun2@mayo.edu}
Xianyang Zhang \email{zhangxiany@stat.tamu.edu}
}
