% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linda.R
\name{linda.wald.test}
\alias{linda.wald.test}
\title{Wald test for bias-corrected regression coefficient}
\usage{
linda.wald.test(
  linda.obj,
  L,
  model = c("LM", "LMM"),
  alpha = 0.05,
  p.adj.method = "BH"
)
}
\arguments{
\item{linda.obj}{return from the \code{linda} function.}

\item{L}{A matrix for testing \code{Lb = 0}, where \code{b} includes the intercept and all fixed effects. Thus the number of columns of
L must be equal to \code{length(variables)+1}, where \code{variables} is from \code{linda.obj}, which does not include the intercept.}

\item{model}{\code{'LM'} or \code{'LMM'} indicating the model fitted in \{linda\} is linear model or linear mixed-effect model.}

\item{alpha}{significance level for testing \code{Lb = 0}.}

\item{p.adj.method}{P-value adjusting approach. See R function \code{p.adjust}. The default is 'BH'.}
}
\value{
A data frame with columns
\item{Fstat}{Wald statistics for each taxon}
\item{df1}{The numerator degrees of freedom}
\item{df2}{The denominator degrees of freedom}
\item{pvalue}{ \code{1 - pf(Fstat, df1, df2)}}
\item{padj}{ \code{p.adjust(pvalue, method = p.adj.method)}}
\item{reject}{ \code{padj <= alpha}}
}
\description{
The function implements Wald test for bias-corrected regression coefficient learned from the \code{linda} function.
One can utilize the function to perform ANOVA-type analyses.
}
\examples{

#install package "phyloseq" for importing "smokers" dataset
ind <- smokers$meta$AIRWAYSITE == 'Throat'
otu.tab <- as.data.frame(smokers$otu[, ind])
meta <- cbind.data.frame(Smoke = factor(smokers$meta$SMOKER[ind]),
                         Sex = factor(smokers$meta$SEX[ind]),
                         Site = factor(smokers$meta$SIDEOFBODY[ind]),
                         SubjectID = factor(smokers$meta$HOST_SUBJECT_ID[ind]))
linda.obj <- linda(otu.tab, meta, formula = '~Smoke+Sex+(1|SubjectID)+(Smoke|Site)', alpha = 0.1,
                   prev.cut = 0.1, lib.cut = 1000, winsor.quan = 0.97)
L <- matrix(c(0, 1, 0, 0, 0, 1), nrow = 2, byrow = TRUE)
result <- linda.wald.test(linda.obj, L, 'LMM', alpha = 0.1)

}
\references{
Huijuan Zhou, Kejun He, Jun Chen, and Xianyang Zhang. LinDA: Linear Models for Differential Abundance
Analysis of Microbiome Compositional Data.
}
\author{
Huijuan Zhou \email{huijuanzhou2019@gmail.com}
Jun Chen \email{Chen.Jun2@mayo.edu}
Xianyang Zhang \email{zhangxiany@stat.tamu.edu}
}
