% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n4_skull_strip.R
\name{n4_skull_strip}
\alias{n4_skull_strip}
\title{Skull Strip image}
\usage{
n4_skull_strip(file, template = system.file("extdata", "pennTemplate",
  "template.nii.gz", package = "LINDA"),
  template_brain = system.file("extdata", "pennTemplate",
  "templateBrain.nii.gz", package = "LINDA"),
  template_mask = system.file("extdata", "pennTemplate",
  "templateBrainMask.nii.gz", package = "LINDA"), verbose = TRUE,
  n_iter = 2)
}
\arguments{
\item{file}{Filename of T1 image}

\item{template}{Template full image}

\item{template_brain}{Template brain image}

\item{template_mask}{Template brain mask}

\item{verbose}{print diagnostic messages}

\item{n_iter}{Number of iterations undergone with \code{\link{abpN4}}
and
\code{\link{abpBrainExtraction}}}
}
\value{
A list of output images, brain and corrected
}
\description{
Skull Strip image
}
