% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_models.R
\docType{data}
\name{rf_model1}
\alias{rf_model1}
\title{Random Forest Models for LINDA}
\format{A \code{randomForest} object}
\usage{
rf_model1
}
\description{
Random Forest Models for LINDA
}
\keyword{datasets}
