% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_prediction.R
\name{run_prediction}
\alias{run_prediction}
\title{Run LINDA Iteration Prediction}
\usage{
run_prediction(img, brain_mask, template_brain, template_mask,
  typeofTransform = "SyN", lesion_mask, reflaxis = 0,
  voxel_resampling = c(2, 2, 2), sigma = 2, verbose = TRUE)
}
\arguments{
\item{img}{image of T1}

\item{brain_mask}{image of brain mask}

\item{template_brain}{image of template brain}

\item{template_mask}{image of template brain mask}

\item{typeofTransform}{type of transformation to perform}

\item{lesion_mask}{image of lesion mask from previous iteration}

\item{reflaxis}{Reflection axis}

\item{voxel_resampling}{Resampling resolution of voxesl}

\item{sigma}{Smoothing factor, passed to
\code{\link{asymmetry_mask}} and
\code{\link{smoothImage}}}

\item{verbose}{Print diagnostic messages}
}
\value{
A list of stuff
}
\description{
Run LINDA Iteration Prediction
}
