% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smokers}
\alias{smokers}
\title{Smoking effect on the human upper respiratory tract}
\format{
A  list with elements
\describe{
  \item{otu}{otu table, 2156 taxa by 290 samples}
  \item{tax}{taxonomy table, 2156 taxa by 7 taxonomic ranks}
  \item{meta}{meta table, 290 samples by 53 sample variables}
  \item{genus}{304 by 290}
  \item{family}{113 by 290}
}
}
\source{
\url{https://qiita.ucsd.edu/} study ID:524
}
\usage{
smokers
}
\description{
A dataset containing "otu", "tax", meta", "genus", family"
Reference: Charlson et al. (2010). Disordered microbial communities in the upper 
respiratory tract of cigarette smokers.
}
\keyword{datasets}
